move temp hbox and vbox funcs directly into demo
